replace interfaces into utils folder, separate images and needName from context into utils
