+++ src/assets/components/AllPlanetsList/AllPlanetsList.tsx
@@ -1,7 +1,7 @@
 import { useContext } from "react";
 import "./AllPlanetsList.scss";
 import PlanetCard from "./PlanetCard";
-import { IPlanet } from "./Interfaces";
+import { IPlanet } from "../../utils/Interfaces";
 import { PlanetsContext } from "../../сontext/PlanetsContext";
 
 export default function AllPlanets() {
