+++ src/assets/utils/Interfaces.ts
