+++ src/assets/utils/images.ts
@@ -1,0 +1,21 @@
+import Earth from "../images/earth.png";
+import Jupiter from "../images/jupiter.png";
+import Mars from "../images/mars.png";
+import Mercury from "../images/mercury.png";
+import Neptune from "../images/neptune.png";
+import Pluto from "../images/pluto.png";
+import Saturn from "../images/saturn.png";
+import Uranus from "../images/uranus.png";
+import Venus from "../images/venus.png";
+
+export const images: { [key: string]: any } = {
+  Mercury: Mercury,
+  Venus: Venus,
+  Earth: Earth,
+  Mars: Mars,
+  Jupiter: Jupiter,
+  Saturn: Saturn,
+  Uranus: Uranus,
+  Neptune: Neptune,
+  Pluto: Pluto,
+};
