+++ src/assets/utils/needNames.ts
@@ -1,0 +1,11 @@
+export const needNames: string[] = [
+  "Mercury",
+  "Venus",
+  "Earth",
+  "Mars",
+  "Jupiter",
+  "Saturn",
+  "Uranus",
+  "Neptune",
+  "Pluto",
+];
