+++ src/assets/сontext/PlanetsContext.tsx
@@ -1,14 +1,7 @@
 import { createContext, useEffect, useState } from "react";
-import { IPlanet } from "../components/AllPlanetsList/Interfaces";
-import Earth from "../images/earth.png";
-import Jupiter from "../images/jupiter.png";
-import Mars from "../images/mars.png";
-import Mercury from "../images/mercury.png";
-import Neptune from "../images/neptune.png";
-import Pluto from "../images/pluto.png";
-import Saturn from "../images/saturn.png";
-import Uranus from "../images/uranus.png";
-import Venus from "../images/venus.png";
+import { IPlanet } from "../utils/Interfaces";
+import { images } from "../utils/images";
+import { needNames } from "../utils/needNames";
 import Error from "../components/Error/Error";
 
 interface IPlanetContext {
@@ -32,30 +25,6 @@ export const PlanetsContextProvider = ({
 }: {
   children: React.ReactNode;
 }) => {
-  const images: { [key: string]: any } = {
-    Mercury: Mercury,
-    Venus: Venus,
-    Earth: Earth,
-    Mars: Mars,
-    Jupiter: Jupiter,
-    Saturn: Saturn,
-    Uranus: Uranus,
-    Neptune: Neptune,
-    Pluto: Pluto,
-  };
-
-  const needNames: string[] = [
-    "Mercury",
-    "Venus",
-    "Earth",
-    "Mars",
-    "Jupiter",
-    "Saturn",
-    "Uranus",
-    "Neptune",
-    "Pluto",
-  ];
-
   const [planets, setPlanets] = useState<IPlanet[]>([]);
   const [loading, setLoading] = useState(false);
   const [error, setError] = useState("");
